Add task 1

+++ 0x0F-function_pointers/function_pointers.h
@@ -2,7 +2,7 @@
 #define function_pointers_h
 
 void print_name(char *name, void (*f)(char *));
-void array_iterator(int *array, size_t size, void (*action)(int));
+void array_iterator(int *array, unsigned int size, void (*action)(int));
 
 
 #endif
